feat: final renames and add aliases
