+++ prowler/providers/microsoft365/services/entra/entra_capolicy_admin_portals_role_limited_access/entra_capolicy_admin_portals_role_limited_access.metadata.json
@@ -2,6 +2,9 @@
   "Provider": "microsoft365",
   "CheckID": "entra_capolicy_admin_portals_role_limited_access",
   "CheckTitle": "Ensure that only administrative roles have access to Microsoft Admin Portals",
+  "CheckAliases": [
+    "entra_admin_portals_role_limited_access"
+  ],
   "CheckType": [],
   "ServiceName": "entra",
   "SubServiceName": "",
