+++ prowler/providers/microsoft365/services/entra/entra_capolicy_administrative_roles_have_mfa_enabled/entra_capolicy_administrative_roles_have_mfa_enabled.metadata.json
@@ -2,6 +2,9 @@
   "Provider": "microsoft365",
   "CheckID": "entra_capolicy_administrative_roles_have_mfa_enabled",
   "CheckTitle": "Ensure multifactor authentication is enabled for all users in administrative roles.",
+  "CheckAliases": [
+    "entra_admin_mfa_enabled_for_administrative_roles"
+  ],
   "CheckType": [],
   "ServiceName": "entra",
   "SubServiceName": "",
