+++ prowler/providers/microsoft365/services/entra/entra_capolicy_ensure_default_user_cannot_create_tenants/entra_capolicy_ensure_default_user_cannot_create_tenants.metadata.json
@@ -2,6 +2,9 @@
   "Provider": "microsoft365",
   "CheckID": "entra_capolicy_ensure_default_user_cannot_create_tenants",
   "CheckTitle": "Ensure that 'Restrict non-admin users from creating tenants' is set to 'Yes'",
+  "CheckAliases": [
+    "entra_ensure_default_user_cannot_create_tenants"
+  ],
   "CheckType": [],
   "ServiceName": "entra",
   "SubServiceName": "",
