--- prowler/providers/microsoft365/services/entra/entra_dynamicgroup_for_guests_created/entra_dynamicgroup_for_guests_created.metadata.json
@@ -1,30 +1,0 @@
-{
-  "Provider": "microsoft365",
-  "CheckID": "entra_dynamicgroup_for_guests_created",
-  "CheckTitle": "Ensure a dynamic group for guest users is created.",
-  "CheckType": [],
-  "ServiceName": "entra",
-  "SubServiceName": "",
-  "ResourceIdTemplate": "",
-  "Severity": "medium",
-  "ResourceType": "Group Settings",
-  "Description": "Ensure that a dynamic group is created for guest users in Microsoft Entra to enforce conditional access policies and security controls automatically.",
-  "Risk": "Without a dynamic group for guest users, administrators may need to manually manage access controls, leading to potential security gaps and inconsistent policy enforcement.",
-  "RelatedUrl": "https://learn.microsoft.com/en-us/entra/identity/users/groups-create-rule",
-  "Remediation": {
-    "Code": {
-      "CLI": "New-MgGroup -DisplayName 'Dynamic Guest Users' -MailNickname 'DynGuestUsers' -MailEnabled $false -SecurityEnabled $true -GroupTypes 'DynamicMembership' -MembershipRule '(user.userType -eq \"Guest\")' -MembershipRuleProcessingState 'On'",
-      "NativeIaC": "",
-      "Other": "1. Navigate to Microsoft Entra admin center https://entra.microsoft.com/. 2. Click to expand Identity > Groups and select All groups. 3. Select 'New group' and configure: Group type: Security, Membership type: Dynamic User. 4. Add dynamic query with rule: (user.userType -eq 'Guest'). 5. Click Save.",
-      "Terraform": ""
-    },
-    "Recommendation": {
-      "Text": "Create a dynamic group for guest users to automate policy enforcement and access control.",
-      "Url": "https://learn.microsoft.com/en-us/entra/identity/users/groups-create-rule"
-    }
-  },
-  "Categories": [],
-  "DependsOn": [],
-  "RelatedTo": [],
-  "Notes": ""
-}
