+++ prowler/providers/microsoft365/services/entra/entra_organization_password_hash_sync_enabled/entra_organization_password_hash_sync_enabled.metadata.json
@@ -2,6 +2,9 @@
   "Provider": "microsoft365",
   "CheckID": "entra_organization_password_hash_sync_enabled",
   "CheckTitle": "Ensure that password hash sync is enabled for hybrid deployments.",
+  "CheckAliases": [
+    "entra_password_hash_sync_enabled"
+  ],
   "CheckType": [],
   "ServiceName": "entra",
   "SubServiceName": "",
