+++ prowler/providers/microsoft365/services/sharepoint/sharepoint_settings_guest_sharing_restricted/sharepoint_settings_guest_sharing_restricted.metadata.json
@@ -1,0 +1,33 @@
+{
+  "Provider": "microsoft365",
+  "CheckID": "sharepoint_settings_guest_sharing_restricted",
+  "CheckTitle": "Ensure that SharePoint guest users cannot share items they don't own.",
+  "CheckAliases": [
+    "sharepoint_guest_sharing_restricted"
+  ],
+  "CheckType": [],
+  "ServiceName": "sharepoint",
+  "SubServiceName": "",
+  "ResourceIdTemplate": "",
+  "Severity": "medium",
+  "ResourceType": "Sharepoint Settings",
+  "Description": "Ensure that guest users in SharePoint cannot share items they do not own, preventing unauthorized disclosure of shared content.",
+  "Risk": "If guest users are allowed to share items they don't own, there is a higher risk of unauthorized data exposure, as external users could share content beyond intended recipients.",
+  "RelatedUrl": "https://learn.microsoft.com/en-us/sharepoint/external-sharing-overview",
+  "Remediation": {
+    "Code": {
+      "CLI": "Set-SPOTenant -PreventExternalUsersFromResharing $True",
+      "NativeIaC": "",
+      "Other": "1. Navigate to SharePoint admin center https://admin.microsoft.com/sharepoint. 2. Click to expand Policies then select Sharing. 3. Expand More external sharing settings and uncheck 'Allow guests to share items they don't own'. 4. Click Save.",
+      "Terraform": ""
+    },
+    "Recommendation": {
+      "Text": "Restrict guest users from sharing items they don't own to enhance security and prevent unauthorized access.",
+      "Url": "https://learn.microsoft.com/en-us/sharepoint/turn-external-sharing-on-or-off"
+    }
+  },
+  "Categories": [],
+  "DependsOn": [],
+  "RelatedTo": [],
+  "Notes": ""
+}
